mvoed mesh creation to assets, fixed up some unnecessary includes
