+++ src/mesh.hpp
@@ -36,12 +36,3 @@ public:
     std::shared_ptr<Buffer> vertexBuffer = nullptr;
     std::shared_ptr<Buffer> indexBuffer = nullptr;
 };
-
-/// @brief Create a mesh object, uploads vertex data to GPU memory.
-/// @param pDeviceContext Render device context to use for mesh creation.
-/// @param vertices 
-/// @param vertexCount 
-/// @param indices 
-/// @param indexCount 
-/// @return A mesh or NULL on error.
-std::shared_ptr<Mesh> createMesh(RenderDeviceContext* pDeviceContext, Vertex* vertices, uint32_t vertexCount, uint32_t* indices, uint32_t indexCount);
